\docType{package}
\name{glmmextra}
\alias{glmmextra}
\alias{glmmextra-package}
\title{glmmextra
Package with useful methods to work with general linear models.
Work mainly with lme4 package, but have method for nlme, too.}
\description{
  glmmextra Package with useful methods to work with
  general linear models. Work mainly with lme4 package, but
  have method for nlme, too.
}

